Move changeset comments controller to feeds module
